requirements.txt
Flask==2.3.3
gunicorn==21.2.0
SQLAlchemy==2.0.21
pg8000==1.29.8
werkzeug==2.3.7
python-dotenv==1.0.0
